requirements.txt
aiohttp
pydub
pyyaml
opuslib
